+++ FlatHouse/DAO.cs
@@ -7,11 +7,10 @@ namespace FlatHouse
 {
     interface DAO<T>
     {
+        public List<T> Get();
+        public T Get(int id);
         public bool Insert(T entity);
         public bool Delete(int id);
         public bool Update(int id, T entity);
-        public List<T> Get();
-        public T Get(int id);
-
     }
 }
