+++ FlatHouse/Flat.cs
@@ -7,24 +7,24 @@ namespace FlatHouse
 {
     public class Flat
     {
-        public int FlatNumber { get; private set; }
+        public int ApartmentNumber { get; private set; }
         public float FloorArea { get; private set; }
         public float Rent { get; private set; }
         public bool IsOccupied { get; private set; }
         public string ResidentName { get; private set; }
-
-        private static int numberCounter = 1;
-
-        public Flat(int flatNumber, float floorArea)
+        public Flat(int apartmentNumber, float floorArea)
         {
-            FlatNumber = flatNumber;
+            ApartmentNumber = apartmentNumber;
             FloorArea = floorArea;
             Rent = floorArea * 100f;
             IsOccupied = false;
         }
-        public Flat(int flatNumber, float floorArea, string residentName) : this(flatNumber, floorArea)
+        public Flat(int apartmentNumber, float floorArea, string residentName) : this(apartmentNumber, floorArea)
         {
-            IsOccupied = true;
+            if (residentName != string.Empty)
+            {
+                IsOccupied = true;
+            }
             ResidentName = residentName;
         }
     }
