+++ FlatHouse/FlatHouse.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MySqlConnector" Version="1.3.1" />
+    <PackageReference Include="MySqlConnector" Version="1.3.8" />
     <PackageReference Include="NEST" Version="7.13.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="5.0.0" />
